+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Version ?? UNKNOWN ??
 
-* The bot can now look up roles with the `!role` command - this requires some setup:
+* **New feature:** `!voteTimer 5 minutes` to start a countdown timer that will perform the `!vote` command after 5 minutes.</br>Times can range from 15 seconds to 20 minutes.</br>Also has corresponding `!stopVoteTimer` command to cancel the timer, as well as shortcuts `!vt` and `!svt`.
+* **New feature:** `The bot can now look up roles with the `!role` command - this requires some setup:
   * `!addset <json url>` adds a custom set to the list for this server
-  * `!removeset <json url>` removes a custom set from the list for this server
+* Added `!setChatChannel <chat channel name>` command to support `!voteTimer`.
 
 ## Version 1.1.0